pyproject.toml
[tool.poetry]
name = "image_composition"
version = "0.1.0"
description = ""
authors = ["watarui <wataruby@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
PyMySql = "^0.9.3"
SQLAlchemy = "^1.3"
flask-sqlalchemy = "^2.4"
Flask-Migrate = "^2.5"
Pillow = "^6.1"
itsdangerous = "^1.1"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest = "^5.0"
isort = "^4.3"
pylint = "^2.3"
rope = "^0.14.0"
flake8 = "^3.7"
mypy = "^0.720.0"
docutils = "^0.15.2"
doc8 = "^0.8.0"
Pygments = "^2.4"
pylint_flask_sqlalchemy = "^0.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
